hum I may have fukkedup
Of course, you need to have all the files needed to build the site using jekyll
in the root directory. And because you can't do anything about it, you can't
change the source, or change the branch/folder, I need to move all my work from
arthur/ to the root.

That's what that commit does. I need to change my strategy with the custom
domain name too.